// packages/backend/src/mxcad/dto/upload-thumbnail-response.dto.d.ts
/**
 * 上传缩略图响应数据 DTO
 */
export declare class UploadThumbnailDataDto {
    fileName: string;
}
/**
 * 上传缩略图响应 DTO
 */
export declare class UploadThumbnailResponseDto {
    code: number;
    message: string;
    data: UploadThumbnailDataDto;
}
//# sourceMappingURL=upload-thumbnail-response.dto.d.ts.map